refactor: move tracing and async dispatchers to dedicated modules
feature/gh34
