+++ ports/api/src/queue_dispatchers/mod.rs
@@ -1,0 +1,5 @@
+mod email_dispatcher;
+mod webhook_dispatcher;
+
+pub(crate) use email_dispatcher::*;
+pub(crate) use webhook_dispatcher::*;
